Address Copilot review: fix circular dep, model/family mismatch, and any cast
