+++ extensions/copilot/src/extension/prompt/vscode-node/requestLoggerImpl.ts
@@ -667,7 +667,15 @@ export class RequestLogger extends AbstractRequestLogger {
 			result.push(`serverRequestId  : ${entry.result.serverRequestId}`);
 		}
 		if (entry.chatParams.body?.tools) {
-			const toolNames = entry.chatParams.body.tools.map((t: any) => isOpenAiFunctionTool(t) ? t.function.name : t.name ?? t.type);
+			const toolNames = entry.chatParams.body.tools.map(t => {
+				if (isOpenAiFunctionTool(t)) {
+					return t.function.name;
+				}
+				if ('name' in t) {
+					return t.name;
+				}
+				return t.type;
+			});
 			const numToolsString = `(${toolNames.length})`;
 			result.push(
 				`<details>`,
